requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pymupdf==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
openpyxl==3.1.5
python-docx==1.1.2
jinja2==3.1.4
pydantic==2.9.2
